added functionality to update due time of notification and moved showNotification function to SharedViewModel

+++ app/src/main/java/com/example/todo/data/models/ToDoData.kt
@@ -14,6 +14,6 @@ data class ToDoData(
     var priority: Priority,
     var description: String,
     var dueTime: Long,
-    var notificationID: Long
+    var notificationID: Int
 ) : Parcelable
